step - extract
