+++ parallel-consumer-core/src/test-integration/java/io/confluent/parallelconsumer/integrationTests/MultiTopicTest.java
@@ -5,15 +5,10 @@
 package io.confluent.parallelconsumer.integrationTests;
 
 import io.confluent.parallelconsumer.ParallelConsumerOptions.ProcessingOrder;
-import io.confluent.parallelconsumer.state.ShardKey;
-import io.confluent.parallelconsumer.state.ShardKey.KeyOrderedKey;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.admin.NewTopic;
-import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.common.TopicPartition;
 import org.hamcrest.Matchers;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 
@@ -21,7 +16,6 @@ import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import static io.confluent.parallelconsumer.ManagedTruth.assertThat;
-import static io.confluent.parallelconsumer.ParallelConsumerOptions.ProcessingOrder.KEY;
 import static one.util.streamex.StreamEx.of;
 import static org.awaitility.Awaitility.await;
 import static org.hamcrest.Matchers.equalTo;
@@ -31,7 +25,6 @@ import static org.hamcrest.Matchers.equalTo;
  */
 @Slf4j
 class MultiTopicTest extends BrokerIntegrationTest<String, String> {
-
 
     @ParameterizedTest
     @EnumSource(ProcessingOrder.class)
@@ -71,29 +64,6 @@ class MultiTopicTest extends BrokerIntegrationTest<String, String> {
         assertThat(getKcu().getLastConsumerConstructed())
                 .hasCommittedToPartition(newTopic)
                 .offset(recordsPerTopic);
-    }
-
-    @Test
-    void keyTest() {
-        ProcessingOrder ordering = KEY;
-        String topicOne = "t1";
-        String keyOne = "k1";
-
-        var reck1 = new ConsumerRecord<>(topicOne, 0, 0, keyOne, "v");
-        ShardKey key1 = ShardKey.of(reck1, ordering);
-        assertThat(key1).isEqualTo(ShardKey.of(reck1, ordering));
-
-        var reck2 = new ConsumerRecord<>(topicOne, 0, 0, "k2", "v");
-        ShardKey of3 = ShardKey.of(reck2, ordering);
-        assertThat(key1).isNotEqualTo(of3);
-
-        var reck3 = new ConsumerRecord<>("t2", 0, 0, keyOne, "v");
-        assertThat(key1).isNotEqualTo(ShardKey.of(reck3, ordering));
-
-        var tp = new TopicPartition(topicOne, 0);
-        KeyOrderedKey keyOrderedKey = new KeyOrderedKey(tp, keyOne);
-        KeyOrderedKey keyOrderedKeyTwo = new KeyOrderedKey(tp, keyOne);
-        assertThat(keyOrderedKey).isEqualTo(keyOrderedKeyTwo);
     }
 
 }
